Centralize SpawnTastyStuff in a single file, decoupling also some include
