+++ src/Beam.cpp
@@ -3,7 +3,6 @@
 
 #include "Beam.h"
 
-#include "CargoBody.h"
 #include "Frame.h"
 #include "Game.h"
 #include "GameSaveError.h"
@@ -12,15 +11,12 @@
 #include "LuaUtils.h"
 #include "Pi.h"
 #include "Planet.h"
-#include "Player.h"
 #include "Sfx.h"
 #include "Ship.h"
 #include "Space.h"
+#include "SpawnTastyStuff.h"
 #include "collider/CollisionContact.h"
 #include "collider/CollisionSpace.h"
-#include "galaxy/Galaxy.h"
-#include "galaxy/StarSystem.h"
-#include "galaxy/SystemBody.h"
 #include "graphics/Graphics.h"
 #include "graphics/Material.h"
 #include "graphics/Renderer.h"
@@ -219,39 +215,6 @@ double Beam::GetRadius() const
 	return sqrt(m_length * m_length);
 }
 
-static void MiningLaserSpawnTastyStuff(Frame *f, const SystemBody *asteroid, const vector3d &pos)
-{
-	lua_State *l = Lua::manager->GetLuaState();
-
-	// lua cant push "const SystemBody", needs to convert to non-const
-	RefCountedPtr<StarSystem> s = Pi::game->GetGalaxy()->GetStarSystem(asteroid->GetPath());
-	SystemBody *liveasteroid = s->GetBodyByPath(asteroid->GetPath());
-
-	// this is an adapted version of "CallMethod", because;
-	// 1, there is no template for LuaObject<LuaTable>::CallMethod(..., SystemBody)
-	// 2, this leaves the return value on the lua stack to be used by "new CargoBody()"
-	LUA_DEBUG_START(l);
-	LuaObject<Player>::PushToLua(Pi::player);
-	lua_pushstring(l, "SpawnMiningContainer");
-	lua_gettable(l, -2);
-	lua_pushvalue(l, -2);
-	lua_remove(l, -3);
-	LuaObject<SystemBody>::PushToLua(liveasteroid);
-	pi_lua_protected_call(l, 2, 1);
-
-	CargoBody *cargo = new CargoBody(LuaRef(l, -1));
-	lua_pop(l, 1);
-	LUA_DEBUG_END(l, 0);
-
-	cargo->SetFrame(f);
-	cargo->SetPosition(pos);
-	const double x = Pi::rng.Double();
-	vector3d dir = pos.Normalized();
-	dir.ArbRotate(vector3d(x, 1 - x, 0), Pi::rng.Double() - .5);
-	cargo->SetVelocity(Pi::rng.Double(100.0, 200.0) * dir);
-	Pi::game->GetSpace()->AddBody(cargo);
-}
-
 void Beam::StaticUpdate(const float timeStep)
 {
 	PROFILE_SCOPED()
@@ -282,14 +245,13 @@ void Beam::StaticUpdate(const float timeStep)
 		// need to test for terrain hit
 		if (GetFrame()->GetBody() && GetFrame()->GetBody()->IsType(Object::PLANET)) {
 			Planet *const planet = static_cast<Planet *>(GetFrame()->GetBody());
-			const SystemBody *b = planet->GetSystemBody();
 			vector3d pos = GetPosition();
 			double terrainHeight = planet->GetTerrainHeight(pos.Normalized());
 			if (terrainHeight > pos.Length()) {
 				// hit the fucker
-				if (b->GetType() == GalaxyEnums::BodyType::TYPE_PLANET_ASTEROID) {
+				if (planet->SystemBodyWrapper::IsType(GalaxyEnums::BodyType::TYPE_PLANET_ASTEROID)) {
 					vector3d n = GetPosition().Normalized();
-					MiningLaserSpawnTastyStuff(planet->GetFrame(), b, n * terrainHeight + 5.0 * n);
+					MiningLaserSpawnTastyStuff(planet->GetFrame(), planet->GetSystemBody(), n * terrainHeight + 5.0 * n);
 					SfxManager::Add(this, TYPE_EXPLOSION);
 				}
 				m_active = false;
