+++ src/SpawnTastyStuff.h
@@ -1,0 +1,14 @@
+// Copyright © 2008-2019 Pioneer Developers. See AUTHORS.txt for details
+// Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
+
+#ifndef SPAWNTASTYSTUFF_H_INCLUDED
+#define SPAWNTASTYSTUFF_H_INCLUDED
+
+#include "vector3.h"
+
+class Frame;
+class SystemBody;
+
+void MiningLaserSpawnTastyStuff(Frame *f, const SystemBody *asteroid, const vector3d &pos);
+
+#endif // SPAWNTASTYSTUFF_H_INCLUDED
